Basic porting/upgrade to vite and modern React, et al versions
