+++ Dockerfile
@@ -11,8 +11,7 @@ FROM nginx
 
 COPY nginx.conf /etc/nginx/conf.d/default.conf
 COPY static /usr/share/nginx/html/static
-COPY --from=builder /src/build/static/css /usr/share/nginx/html/static/css
-COPY --from=builder /src/build/static/js /usr/share/nginx/html/static/js
-COPY --from=builder /src/build/index.html /usr/share/nginx/html/index.html
+COPY --from=builder /src/dist/assets /usr/share/nginx/html/assets
+COPY --from=builder /src/dist/index.html /usr/share/nginx/html/index.html
 EXPOSE 80
