+++ Dockerfile.builder
@@ -1,5 +1,4 @@
-FROM node:8-alpine as builder
+FROM node:16-alpine as builder
 WORKDIR /src
 COPY package.json /src
 RUN npm install
-RUN npm rebuild node-sass
