+++ index.html
@@ -1,0 +1,16 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <link rel="icon" type="image/svg+xml" href="/vite.svg" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>New & Improved DMS Frontend</title>
+
+    <link href="/static/font-awesome/css/all.min.css" rel="stylesheet" type="text/css" />
+    <link rel="icon" type="image/x-icon" href="/static/images/favicon.ico">
+  </head>
+  <body>
+    <div id="root"></div>
+    <script type="module" src="/src/index.jsx"></script>
+  </body>
+</html>
