+++ package.json
@@ -1,41 +1,34 @@
 {
-  "name": "webapp",
-  "version": "0.1.0",
+  "name": "dmsfront-rereact",
   "private": true,
-  "proxy": "http://dm:8010",
-  "dependencies": {
-    "@canonical/react-components": "^0.24.0",
-    "@emotion/react": "^11.7.0",
-    "@emotion/styled": "^11.6.0",
-    "@mui/material": "^5.2.2",
-    "axios": "^0.24.0",
-    "history": "^4.7.2",
-    "jwt-decode": "^2.2.0",
-    "moment": "^2.29.1",
-    "moment-timezone": "^0.5.33",
-    "react": "^16.6.3",
-    "react-dom": "^16.6.3",
-    "react-highlight": "^0.14.0",
-    "react-scripts": "^3.4.3",
-    "react-tooltip": "^4.2.21",
-    "vanilla-framework": "^2.37.1"
-  },
+  "version": "0.0.0",
+  "type": "module",
   "scripts": {
-    "build-css": "node-sass src/ -o src/",
-    "start": "npm run build-css && react-scripts start",
-    "buildonly": "react-scripts build",
-    "build": "npm run build-css && react-scripts build",
-    "test": "npm run build-css && react-scripts test --env=jsdom",
-    "eject": "react-scripts eject"
+    "dev": "vite",
+    "build": "tsc && vite build",
+    "preview": "vite preview"
+  },
+  "dependencies": {
+    "@canonical/react-components": "^0.38.0",
+    "@emotion/react": "^11.10.6",
+    "@emotion/styled": "^11.10.6",
+    "@mui/material": "^5.11.12",
+    "axios": "^1.3.4",
+    "filesize": "^10.0.6",
+    "history": "^5.3.0",
+    "jwt-decode": "^3.1.2",
+    "moment": "^2.29.4",
+    "moment-timezone": "^0.5.41",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "vanilla-framework": "3.11.0"
   },
   "devDependencies": {
-    "filesize": "5.0.3",
-    "node-sass": "^4.14.1"
-  },
-  "browserslist": [
-    ">0.2%",
-    "not dead",
-    "not ie <= 11",
-    "not op_mini all"
-  ]
+    "@types/react": "^18.0.27",
+    "@types/react-dom": "^18.0.10",
+    "@vitejs/plugin-react-swc": "^3.0.0",
+    "sass": "^1.58.3",
+    "typescript": "^4.9.3",
+    "vite": "^4.1.0"
+  }
 }
