+++ src/App.test.js
@@ -17,7 +17,7 @@
 
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './App';
+import App from './App.jsx';
 
 it('renders without crashing', () => {
   const div = document.createElement('div');
