+++ src/components/Device.jsx
