+++ src/components/DeviceSnaps.jsx
