+++ src/index.jsx
@@ -17,7 +17,7 @@
 
 import React from 'react';
 import ReactDOM from 'react-dom';
-import App from './App';
+import App from './App.jsx';
 import {getAuthToken}      from './components/Utils'
 
 getAuthToken( (token) => {
