+++ src/main.tsx
@@ -1,0 +1,10 @@
+import * as React from 'react'
+import * as ReactDOM from 'react-dom/client'
+import App from './App'
+// import './index.css'
+
+ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
+  <React.StrictMode>
+    <App />
+  </React.StrictMode>,
+)
